feat(shared): Move Twitch types to constants module
