+++ apps/discord-bot-internal/src/services/streams.ts
@@ -7,7 +7,7 @@ import {
 import { config } from '../config';
 import { Service } from '../types/service';
 import { TwitchAPI } from '../twitch-api';
-import { TwitchStream } from '../types/twitch';
+import { TwitchStream } from '@momentum/constants';
 import { sanitizeMarkdown, timeSpanToPrettyPrint } from '../utils';
 import { logger } from '../logger';
