+++ libs/constants/src/types/models/models.ts
@@ -543,10 +543,37 @@ export interface CompletedRun {
 //#region Misc
 
 export interface TwitchStream {
-  title: string;
+  id: string;
+  user_id: string;
+  user_login: string;
   user_name: string;
+  game_id: string;
+  game_name: string;
+  type: string;
+  title: string;
+  tags: string[];
   viewer_count: number;
   started_at: string;
+  language: string;
   thumbnail_url: string;
+  /** @deprecated */
+  tag_ids: any[];
+  is_mature: boolean;
 }
+
+export interface TwitchUser {
+  id: string;
+  login: string;
+  display_name: string;
+  type: string;
+  broadcaster_type: string;
+  description: string;
+  profile_image_url: string;
+  offline_image_url: string;
+  /** @deprecated */
+  view_count: number;
+  email: string;
+  created_at: string;
+}
+
 //#endregion
